Moved files to standard directory structure
