+++ lib/discovery-indexer/errors.rb
