+++ lib/discovery-indexer/logging.rb
