+++ lib/discovery-indexer/mapper/general_mapper.rb
