+++ lib/discovery-indexer/reader/modsxml.rb
