+++ lib/discovery-indexer/reader/modsxml_reader.rb
