+++ lib/discovery-indexer/reader/purlxml.rb
