+++ lib/discovery-indexer/reader/purlxml_model.rb
