+++ lib/discovery-indexer/reader/purlxml_parser.rb
