+++ lib/discovery-indexer/reader/purlxml_parser_strict.rb
