+++ lib/discovery-indexer/reader/purlxml_reader.rb
