+++ lib/discovery-indexer/version.rb
