+++ lib/discovery-indexer/writer/solr_client.rb
