+++ lib/discovery-indexer/writer/solr_writer.rb
