+++ spec/discovery-indexer/mapper/general_mapper_spec.rb
