+++ spec/discovery-indexer/reader/modsxml_reader_spec.rb
