+++ spec/discovery-indexer/reader/purlxml_parser_strict_spec.rb
