+++ spec/discovery-indexer/reader/purlxml_reader_spec.rb
