+++ spec/discovery-indexer/reader/purlxml_spec.rb
