+++ spec/discovery-indexer/writer/solr_client_spec.rb
