+++ spec/discovery-indexer/writer/solr_writer_spec.rb
